Organize files to separate ng1/ng2 stuff. Skeleton classes for ng2.
