+++ bootstrapper/tabs/tabsNg1.html
@@ -1,0 +1,16 @@
+<h3>Tabs</h3>
+<label>Tabset</label>
+<rl-tabset>
+	<rl-tab>
+		<rl-tab-header>Header</rl-tab-header>
+		<rl-tab-content>Content 1</rl-tab-content>
+		<rl-tab-footer>Footer 1</rl-tab-footer>
+	</rl-tab>
+	<rl-tab>
+		<rl-tab-header>Header 2</rl-tab-header>
+		<rl-tab-content>Content 2</rl-tab-content>
+		<rl-tab-footer>Footer 2</rl-tab-footer>
+	</rl-tab>
+</rl-tabset>
+<label>Multi-step indicator</label>
+<rl-multi-step-indicator steps="tabs.steps"></rl-multi-step-indicator>
