+++ source/components/tabs/tab.html
@@ -1,10 +1,10 @@
-<div class="tab-pane" ng-class="{ 'active': tab.header.isVisible }">
+<div class="tab-pane" ng-class="{ 'active': header.isVisible }">
 	<div class="tab-body">
-		<div ng-transclude="contentSlot" ng-form="tabForm"></div>
+        <ng-content select="tab-content"></ng-content>
 		<div class="clearfix"></div>
 	</div>
 	<div class="tab-footer">
-		<div ng-transclude="footerSlot"></div>
+		<ng-content select="tab-footer"></ng-content>
 		<div class="clearfix"></div>
 	</div>
 </div>
