+++ source/components/tabs/tab.ng1.html
@@ -1,0 +1,10 @@
+<div class="tab-pane" ng-class="{ 'active': tab.header.isVisible }">
+	<div class="tab-body">
+		<div ng-transclude="contentSlot" ng-form="tabForm"></div>
+		<div class="clearfix"></div>
+	</div>
+	<div class="tab-footer">
+		<div ng-transclude="footerSlot"></div>
+		<div class="clearfix"></div>
+	</div>
+</div>
