+++ source/components/tabs/tabset.html
@@ -1,11 +1,11 @@
 <div class="nav-tabs-dropshadow">
 	<ul class="nav nav-tabs">
-		<li ng-repeat="tabHeader in tabset.tabHeaders" ng-click="tabset.select(tabHeader)"
+		<li *ngFor="let tabHeader of tabHeaders" (click)="select(tabHeader)"
 			ng-class="{ 'active': tabHeader.isVisible, 'error': !tabHeader.isValid }">
-			<a><span ng-bind-html="tabHeader.template"></span></a>
+			<a><span [innerHTML]="tabHeader.template"></span></a>
 		</li>
 	</ul>
 	<div class="tab-content">
-		<div ng-transclude></div>
+		<ng-content></ng-content>
 	</div>
 </div>
